ERMS-1694: Pakete gezielt entknüpfen für Teilnehmer als Bulk-Version

+++ app/grails-app/controllers/com/k_int/kbplus/SubscriptionController.groovy
@@ -1709,41 +1709,43 @@ class SubscriptionController extends AbstractDebugController {
         result.parentPackages = result.parentSub.packages.sort { it.pkg.name }
 
         def validSubChilds = Subscription.findAllByInstanceOf(result.parentSub)
-
         validSubChilds.each { subChild ->
-
             subChild.packages.pkg.each { pkg ->
-                if(!CostItem.executeQuery('select ci from CostItem ci where ci.subPkg.pkg = :sp',[sp:pkg])) {
-                    def query = "from IssueEntitlement ie, Package pkg where ie.subscription =:sub and pkg.id =:pkg_id and ie.tipp in ( select tipp from TitleInstancePackagePlatform tipp where tipp.pkg.id = :pkg_id ) "
-                    def queryParams = [sub: subChild, pkg_id: pkg.id]
+
+                def pkg_to_unlink = params.package_All ? SubscriptionPackage.get(params.package_All).pkg : null
+
+                if(pkg_to_unlink == null || pkg_to_unlink == pkg) {
+
+                    if (!CostItem.executeQuery('select ci from CostItem ci where ci.subPkg.pkg = :sp', [sp: pkg])) {
+                        def query = "from IssueEntitlement ie, Package pkg where ie.subscription =:sub and pkg.id =:pkg_id and ie.tipp in ( select tipp from TitleInstancePackagePlatform tipp where tipp.pkg.id = :pkg_id ) "
+                        def queryParams = [sub: subChild, pkg_id: pkg.id]
 
 
+                        if (subChild.isEditableBy(result.user)) {
+                            result.editable = true
+                            if (params.withIE) {
+                                //delete matches
+                                IssueEntitlement.withTransaction { status ->
+                                    removePackagePendingChanges(pkg.id, subChild.id, params.withIE)
+                                    def deleteIdList = IssueEntitlement.executeQuery("select ie.id ${query}", queryParams)
 
-                    if (subChild.isEditableBy(result.user)) {
-                        result.editable = true
-                        if (params.withIE) {
-                            //delete matches
-                            IssueEntitlement.withTransaction { status ->
-                                removePackagePendingChanges(pkg.id, subChild.id, params.withIE)
-                                def deleteIdList = IssueEntitlement.executeQuery("select ie.id ${query}", queryParams)
+                                    if (deleteIdList) {
+                                        IssueEntitlementCoverage.executeUpdate("delete from IssueEntitlementCoverage ieCov where ieCov.issueEntitlement.id in (:delList)", [delList: deleteIdList])
+                                        PriceItem.executeUpdate("delete from PriceItem pi where pi.issueEntitlement.id in (:delList)", [delList: deleteIdList])
+                                        IssueEntitlement.executeUpdate("delete from IssueEntitlement ie where ie.id in (:delList)", [delList: deleteIdList])
+                                    }
+                                    SubscriptionPackage.executeUpdate("delete from SubscriptionPackage sp where sp.pkg=? and sp.subscription=? ", [pkg, subChild])
 
-                                if (deleteIdList) {
-                                    IssueEntitlementCoverage.executeUpdate("delete from IssueEntitlementCoverage ieCov where ieCov.issueEntitlement.id in (:delList)", [delList: deleteIdList])
-                                    PriceItem.executeUpdate("delete from PriceItem pi where pi.issueEntitlement.id in (:delList)",[delList: deleteIdList])
-                                    IssueEntitlement.executeUpdate("delete from IssueEntitlement ie where ie.id in (:delList)", [delList: deleteIdList])
+                                    flash.message = message(code: 'subscription.linkPackagesMembers.unlinkInfo.withIE.successful')
                                 }
+                            } else {
                                 SubscriptionPackage.executeUpdate("delete from SubscriptionPackage sp where sp.pkg=? and sp.subscription=? ", [pkg, subChild])
 
-                                flash.message = message(code: 'subscription.linkPackagesMembers.unlinkInfo.withIE.successful')
+                                flash.message = message(code: 'subscription.linkPackagesMembers.unlinkInfo.onlyPackage.successful')
                             }
-                        } else {
-                            SubscriptionPackage.executeUpdate("delete from SubscriptionPackage sp where sp.pkg=? and sp.subscription=? ", [pkg, subChild])
-
-                            flash.message = message(code: 'subscription.linkPackagesMembers.unlinkInfo.onlyPackage.successful')
                         }
-                    }
+                    } else flash.error += "Für das Paket ${pkg.name} von ${subChild.getSubscriber().name} waren noch Kosten anhängig. Das Paket wurde daher nicht entknüpft."
                 }
-                else flash.error += "Für das Paket ${pkg.name} von ${subChild.getSubscriber().name} waren noch Kosten anhängig. Das Paket wurde daher nicht entknüpft."
             }
         }
 
@@ -3117,7 +3119,7 @@ class SubscriptionController extends AbstractDebugController {
                     }
                 }
 
-                log.debug("Global Record Source URL: " +gri.source.baseUrl)
+                log.debug("linkPackage. Global Record Source URL: " +gri.source.baseUrl)
                 //if(Package.findByGokbId(grt.owner.uuid)) {
                 String addType = params.addType
                     executorWrapperService.processClosure({
